feat: refractor with bin and add checkProjectExists
